Moved GameInputHandler states into their own files/classes.
