+++ src/Engine/GameInput/GameInputHandler.java
@@ -1,22 +1,13 @@
 package Engine.GameInput;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Stack;
 
 import CommandingOfficers.Commander;
 import Engine.GameAction;
-import Engine.GameActionSet;
 import Engine.Path;
-import Engine.Utils;
 import Engine.XYCoord;
-import Engine.GameAction.UnitProductionAction;
-import Engine.GameAction.UnloadAction;
 import Terrain.GameMap;
-import Terrain.Location;
-import Units.Unit;
-import Units.UnitModel;
 
 /************************************************************
  * Handles converting user input into game actions.
@@ -176,84 +167,7 @@ public class GameInputHandler
   /** @return The current set of coordinates from which to choose, or null we are not selecting a location. */
   public ArrayList<XYCoord> getCoordinateOptions()
   {
-    return peekCurrentState().getOptions().myCoords;
-  }
-
-  /************************************************************
-   *  Just a struct class to hold information
-   *  for easy sharing across State objects.
-   ************************************************************/
-  private class StateData
-  {
-    public final GameMap gameMap;
-    public final Commander commander;
-    public Unit unitActor = null;
-    public GameActionSet actionSet = null;
-    public Path path = null;
-    public ArrayList<? extends Object> menuOptions = null; // Just require a toString().
-    public Map<Unit, XYCoord> unitLocationMap = null; // Used to map units to unload locations.
-    public StateData(GameMap map, Commander co)
-    {
-      gameMap = map;
-      commander = co;
-    }
-  }
-
-  /************************************************************
-   * Provides access to the valid options for the current state
-   * via its accessors. If an accessor returns null, that input
-   * type is not supported in this state.
-   * The appropriate accessor can be determined by checking for
-   * null or by inspecting inputMode.
-   * If inputMode==PATH_SELECT or CONSTRAINED_TILE_SELECT,
-   *   then use getCoordinateOptions(). 
-   * If inputMode==MENU_SELECT, use getMenuOptions().
-   * If inputMode==ACTION_READY, use (and execute) getAction(). 
-   ************************************************************/
-  private static class OptionSet
-  {
-    public final InputMode inputMode;
-    private ArrayList<XYCoord> myCoords = null;
-    private Object[] myMenuOptions = null;
-    private GameAction myAction = null;
-
-    public OptionSet()
-    {
-      inputMode = InputMode.FREE_TILE_SELECT;
-    }
-
-    public OptionSet(InputMode mode, ArrayList<XYCoord> coords)
-    {
-      inputMode = mode;
-      myCoords = coords;
-    }
-
-    public OptionSet(Object[] menuOptions)
-    {
-      inputMode = InputMode.MENU_SELECT;
-      myMenuOptions = menuOptions;
-    }
-
-    public OptionSet(GameAction action)
-    {
-      inputMode = InputMode.ACTION_READY;
-      myAction = action;
-    }
-
-    public Object[] getMenuOptions()
-    {
-      return myMenuOptions;
-    }
-
-    public ArrayList<XYCoord> getCoordinateOptions()
-    {
-      return myCoords;
-    }
-
-    public GameAction getAction()
-    {
-      return myAction;
-    }
+    return peekCurrentState().getOptions().getCoordinateOptions();
   }
 
   /************************************************************
@@ -262,498 +176,5 @@ public class GameInputHandler
   public static interface StateChangedCallback
   {
     public void onStateChange();
-  }
-
-  /************************************************************
-   * Abstract base class for all input states.                *
-   ************************************************************/
-  private static abstract class GameInputState
-  {
-    /** The current GameActionBuilder state data,
-     *  shared with all State instances. */
-    protected final StateData myStateData;
-    protected final OptionSet myOptions;
-
-    public GameInputState(StateData data)
-    {
-      myStateData = data;
-      myOptions = initOptions();
-    }
-
-    /** Forces each subclass to create an OptionSet.
-     *  You shouldn't need anything but the contents
-     *  of myStateData to do this. If you want high-
-     *  lighted map tiles in this state, then handle
-     *  that here too. */
-    protected abstract OptionSet initOptions();
-
-    /** Get an object representing the valid selections that may be
-     *  made in the current state. */
-    public OptionSet getOptions()
-    {
-      return myOptions;
-    }
-
-    // Default implementations of select() will just keep us
-    // in the same state. Subclasses will define transitions.
-    public GameInputState select(Path path)
-    {
-      return null;
-    }
-    public GameInputState select(Object option)
-    {
-      return this;
-    }
-    public GameInputState select(XYCoord coord)
-    {
-      return this;
-    }
-    /** Undo any StateData changes. */
-    public void back(){}
-  }
-
-  /************************************************************
-   * State for before any user input has been received, or    *
-   * immediately after reset() has been called.               *
-   ************************************************************/
-  private static class DefaultState extends GameInputState
-  {
-    public DefaultState(StateData data)
-    {
-      super(data);
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      // Use the default OptionSet constructor, which allows free tile selection.
-      return new OptionSet();
-    }
-
-    @Override
-    public GameInputState select(XYCoord coord)
-    {
-      GameInputState next = this;
-      Location loc = myStateData.gameMap.getLocation(coord);
-      Unit resident = loc.getResident();
-      if( null != resident && resident.CO == myStateData.commander )
-      {
-        next = select(resident);
-      }
-      else if( (loc.getOwner() == myStateData.commander) && myStateData.commander.getShoppingList(loc.getEnvironment().terrainType).size() > 0 )
-      {
-        ArrayList<UnitModel> buildables = myStateData.commander.getShoppingList(loc.getEnvironment().terrainType);
-        myStateData.menuOptions = buildables;
-        next = new SelectUnitProduction(myStateData, buildables, coord);
-      }
-      return next;
-    }
-
-    public GameInputState select(Unit unit)
-    {
-      GameInputState next = this;
-      if( unit != null
-          && (!unit.isTurnOver    // If it's our unit and it's ready to go.
-              // || unit.CO != myStateData.commander // Also allow checking the move radius of others' units.
-              ) )
-      {
-        myStateData.unitActor = unit;
-        next = new SelectMoveLocation(myStateData);
-      }
-      else
-      {
-        System.out.println("Invalid unit selected.");
-        if( unit == null ) System.out.println("  Unit is null.");
-        if( unit.isTurnOver ) System.out.println("  Unit has already moved.");
-      }
-      return next;
-    }
-  }
-
-  /************************************************************
-   * Presents options for building a unit.                    *
-   ************************************************************/
-  private static class SelectUnitProduction extends GameInputState
-  {
-    private ArrayList<UnitModel> myUnitModels = null;
-    private XYCoord myProductionLocation = null;
-
-    public SelectUnitProduction(StateData data, ArrayList<UnitModel> buildables, XYCoord buildLocation)
-    {
-      super(data);
-      myUnitModels = buildables;
-      myProductionLocation = buildLocation;
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      OptionSet options = null;
-      if( null != myStateData.menuOptions )
-      {
-        String[] modelStrings = new String[myStateData.menuOptions.size()];
-        for( int i = 0; i < myStateData.menuOptions.size(); ++i )
-        {
-          modelStrings[i] = myStateData.menuOptions.get(i).toString();
-        }
-        options = new OptionSet(modelStrings);
-      }
-      return options;
-    }
-
-    @Override
-    public GameInputState select(Object option)
-    {
-      GameInputState next = this;
-
-      if( null != option && null != myUnitModels )
-      {
-        for( UnitModel model : myUnitModels )
-        {
-          if( option.equals(model.toString()))
-          {
-            myStateData.actionSet = new GameActionSet(new GameAction.UnitProductionAction(myStateData.gameMap, myStateData.commander, model, myProductionLocation), false);
-            next = new ActionReady(myStateData);
-          }
-        }
-      }
-
-      return next;
-    }
-  }
-
-  /************************************************************
-   * State to allow choosing a unit's path.                   *
-   ************************************************************/
-  private static class SelectMoveLocation extends GameInputState
-  {
-    public SelectMoveLocation(StateData data)
-    {
-      super(data);
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      // Get valid move locations and highlight the relevant map tiles.
-      ArrayList<XYCoord> moveLocations = Utils.findPossibleDestinations(myStateData.unitActor, myStateData.gameMap);
-      for( XYCoord xy : moveLocations )
-      {
-        myStateData.gameMap.getLocation(xy.xCoord, xy.yCoord).setHighlight(true);
-      }
-      // Build and return our OptionSet.
-      return new OptionSet(InputMode.PATH_SELECT, moveLocations);
-    }
-
-    @Override
-    public GameInputState select(Path path)
-    {
-      GameInputState next = this;
-      if( myOptions.getCoordinateOptions().contains(new XYCoord(path.getEnd().x, path.getEnd().y))
-          && Utils.isPathValid(myStateData.unitActor, path, myStateData.gameMap) )
-      {
-        // The path ends on a valid move location, and is traversable by the unit. Store it.
-        myStateData.path = path;
-
-        // We'll be changing state, so unset the map highlights.
-        for( XYCoord xy : myOptions.getCoordinateOptions() )
-        {
-          myStateData.gameMap.getLocation(xy.xCoord, xy.yCoord).setHighlight(false);
-        }
-
-        // Construct the next state instance.
-        next = new SelectUnitAction(myStateData);
-      }
-      return next;
-    }
-
-    @Override
-    public void back()
-    {
-      myStateData.path = null;
-    }
-  }
-
-  /************************************************************
-   * State to allow selecting an action for a unit.           *
-   ************************************************************/
-  private static class SelectUnitAction extends GameInputState
-  {
-    // Don't provide a default value, or it will override the value
-    // set by the call to initOptions() in the super-constructor.
-    private ArrayList<GameActionSet> myUnitActions;
-
-    public SelectUnitAction(StateData data)
-    {
-      super(data);
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      // Get the set of actions this unit could perform from the target location.
-      myUnitActions = myStateData.unitActor.getPossibleActions(myStateData.gameMap, myStateData.path);
-
-      // Create our new menu options.
-      String[] options = new String[myUnitActions.size()];
-      for(int i = 0; i < myUnitActions.size(); ++i)
-      {
-        options[i] = myUnitActions.get(i).toString();
-      }
-      return new OptionSet(options);
-    }
-
-    @Override
-    public GameInputState select(Object menuOption)
-    {
-      GameInputState next = this;
-      GameActionSet chosenSet = null;
-      if( null != menuOption )
-      {
-        for(GameActionSet set : myUnitActions)
-        {
-          if( set.toString().equals(menuOption) )
-          {
-            chosenSet = set;
-            break;
-          }
-        }
-      }
-
-      if( null != chosenSet )
-      {
-        // Add the action set to our state data.
-        myStateData.actionSet = chosenSet;
-
-        // If the action type requires no target, then we are done collecting input.
-        if( !chosenSet.isTargetRequired() )
-        {
-          next = new ActionReady(myStateData);
-        }
-        else
-        {
-          // We need more input before an action is ready; What kind of input depends on the type of action.
-          switch(chosenSet.getSelected().getType())
-          {
-            case ATTACK:
-              // We need to select a target to attack.
-              next = new SelectActionTarget(myStateData);
-              break;
-            case UNLOAD:
-              // We need to select a unit to unload.
-              next = new SelectCargo(myStateData);
-              break;
-          }
-        }
-      }
-      return next;
-    }
-
-    @Override
-    public void back()
-    {
-      myStateData.actionSet = null;
-    }
-  }
-
-  /************************************************************
-   * Allows selecting an action's target.                     *
-   ************************************************************/
-  private static class SelectActionTarget extends GameInputState
-  {
-    public SelectActionTarget(StateData data)
-    {
-      super(data);
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      // Set the target-location highlights.
-      myStateData.gameMap.clearAllHighlights();
-      ArrayList<XYCoord> targets = myStateData.actionSet.getTargetedLocations();
-      for( XYCoord targ : targets )
-      {
-        myStateData.gameMap.getLocation(targ).setHighlight(true);
-      }
-
-      // We can only attack the selected tiles, and they may be disjoint, so use constrained tile select.
-      return new OptionSet(InputMode.CONSTRAINED_TILE_SELECT, targets);
-    }
-
-    @Override
-    public GameInputState select(XYCoord targetLocation)
-    {
-      GameInputState next = this;
-
-      // Find the action that this target location belongs to.
-      // By virtue of the fact that GameActionSets should be homogenous, and it should be
-      // nonsensical to have two actions of the same type targeting the same location, this
-      // should be enough to uniquely identify the action we want.
-      // Well OK, UNLOAD action sets can have different actions with the same target location.
-      //   Fortunately, we don't handle those here.
-      for( int i = 0; i < myStateData.actionSet.getTargetedLocations().size(); ++i )
-      {
-        // If the selected target location corresponds to this action, keep it selected.
-        if( targetLocation.equals(myStateData.actionSet.getSelected().getTargetLocation()) )
-        {
-          // ActionReady will just choose whatever action is selected to perform, so no changes here.
-          next = new ActionReady(myStateData);
-          break;
-        }
-
-        // That wasn't it; move to the next one.
-        myStateData.actionSet.next();
-      }
-
-      // If we couldn't find an action for this location, return the current state.
-      return next;
-    }
-  }
-
-  /************************************************************
-   * State to choose which Unit will be kicked off the bus.   *
-   ************************************************************/
-  private static class SelectCargo extends GameInputState
-  {
-    public SelectCargo(StateData data)
-    {
-      super(data);
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      // Collect the names of all units held by unitActor.
-      ArrayList<Object> cargoes = new ArrayList<Object>();
-      for( int i = 0; i < myStateData.unitActor.heldUnits.size(); ++i )
-      {
-        // Don't include the unit if it's already set to be unloaded.
-        if( (null == myStateData.unitLocationMap) || !myStateData.unitLocationMap.containsKey(myStateData.unitActor.heldUnits.get(i)) )
-        {
-          cargoes.add( myStateData.unitActor.heldUnits.get(i) );
-        }
-      }
-      if( null != myStateData.unitLocationMap && myStateData.unitLocationMap.size() > 0 )
-      {
-        // If at least one unit is being offboarded already, offer a "DONE" option so we aren't forced to unload everything.
-        cargoes.add("DONE");
-      }
-      Object[] cargoArray = new Object[cargoes.size()];
-      cargoes.toArray(cargoArray);
-      return new OptionSet(cargoArray);
-    }
-
-    @Override
-    public GameInputState select(Object option)
-    {
-      GameInputState next = this;
-
-      // Add a unitLocationMap to our state if we don't have one already.
-      if( null == myStateData.unitLocationMap )
-      {
-        myStateData.unitLocationMap = new HashMap<Unit, XYCoord>();
-      }
-
-      if( option.equals("DONE") )
-      {
-        // Since we don't want to drop any additional units, build the GameAction and move to ActionReady.
-        GameAction ga = new GameAction.UnloadAction(myStateData.gameMap, myStateData.unitActor, myStateData.path, myStateData.unitLocationMap);
-
-        // Override the current ActionSet with a new one, since we just redefined it.
-        myStateData.actionSet = new GameActionSet( ga, true );
-        next = new ActionReady(myStateData);
-      }
-      for( Unit cargo : myStateData.unitActor.heldUnits )
-      {
-        if(cargo == option)
-        {
-          next = new SelectCargoDropLocation(myStateData, cargo);
-        }
-      }
-
-      return next;
-    }
-
-    @Override
-    public void back()
-    {
-      myStateData.unitLocationMap = null;
-    }
-  }
-
-  /************************************************************
-   * State to choose where to drop the unit.                  *
-   ************************************************************/
-  private static class SelectCargoDropLocation extends GameInputState
-  {
-    Unit myCargo = null;
-
-    public SelectCargoDropLocation(StateData data, Unit newDrop)
-    {
-      super(data);
-      myCargo = newDrop;
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      ArrayList<XYCoord> dropoffLocations = myStateData.actionSet.getTargetedLocations();
-      dropoffLocations.removeAll(myStateData.unitLocationMap.values()); // Remove any drop locations that are already reserved.
-      return new OptionSet(InputMode.CONSTRAINED_TILE_SELECT, dropoffLocations);
-    }
-
-    @Override
-    public GameInputState select(XYCoord location)
-    {
-      GameInputState next = this;
-
-      if( myStateData.actionSet.getTargetedLocations().contains(location) )
-      {
-        // Add the new dropoff to myStateData.
-        myStateData.unitLocationMap.put(myCargo, location);
-
-        // If we have the ability to unload another unit as well, go forward to SelectCargo.
-        if( myStateData.unitActor.heldUnits.size() > myStateData.unitLocationMap.size() // More units we could unload.
-            && getOptions().getCoordinateOptions().size() > myStateData.unitLocationMap.size() ) // More spaces to drop them in.
-        {
-          next = new SelectCargo(myStateData);
-        }
-        else
-        {
-          // Since we can't drop any additional units, build the GameAction and move to ActionReady.
-          GameAction ga = new GameAction.UnloadAction(myStateData.gameMap, myStateData.unitActor, myStateData.path, myStateData.unitLocationMap);
-
-          // Override the current ActionSet with a new one, since we just redefined it.
-          myStateData.actionSet = new GameActionSet( ga, true );
-          next = new ActionReady(myStateData);
-        }
-      }
-
-      return next;
-    }
-
-    @Override
-    public void back()
-    {
-      myStateData.unitLocationMap.remove(myCargo);
-    }
-  }
-
-  /************************************************************
-   * Terminal state - just provides the selected action.      *
-   ************************************************************/
-  private static class ActionReady extends GameInputState
-  {
-    public ActionReady(StateData data)
-    {
-      super(data);
-    }
-
-    @Override
-    protected OptionSet initOptions()
-    {
-      return new OptionSet(myStateData.actionSet.getSelected());
-    }
   }
 }
